\name{check_data}
\alias{check_data}
\title{Check Data}
\usage{
  check_data(Data)
}
\arguments{
  \item{Data}{numeric matrix or data frame containing the
  manifest variables.}
}
\value{
  validated Data
}
\description{
  Internal function not to be called by the user
}
\details{
  Internal function. \code{check_data} is called by
  \code{check_args}.
}
\section{Warning}{
  Do NOT use this function unless you are ME, a package
  developer, or a jedi user who really knows what is doing
  (seriously!)
}
\keyword{internal}

