\name{check_scheme}
\alias{check_scheme}
\title{Check scheme}
\usage{
  check_scheme(scheme)
}
\arguments{
  \item{scheme}{character indicating the inner weighting
  scheme}
}
\value{
  validated scheme
}
\description{
  Internal function not to be called by the user
}
\details{
  Internal function. \code{check_scheme} is called by
  \code{check_specs}.
}
\section{Warning}{
  Do NOT use this function unless you are ME, a package
  developer, or a jedi user who really knows what is doing
  (seriously!)
}
\keyword{internal}

