\name{dummy.G}
\alias{dummy.G}
\title{Dummy by Giorgio}
\usage{
  dummy.G(Y, X)
}
\arguments{
  \item{Y}{a matrix}

  \item{X}{a matrix}
}
\value{
  list with Xdummy, Quant, eta2
}
\description{
  Internal function not to be called by the user
}
\details{
  Internal function. \code{dummy.G} is called by
  \code{get_PLSRdoubleQ}.
}
\section{Warning}{
  Do NOT use this function unless you are ME, a package
  developer, or a jedi user who really knows what is doing
  (seriously!)
}
\keyword{internal}

