\name{get_effects}
\alias{get_effects}
\title{Path coefficient effects for \code{plspm}}
\usage{
  get_effects(Path)
}
\arguments{
  \item{Path}{matrix of path coefficients}
}
\value{
  data frame with relationships, direct, indirect, and
  total effects
}
\description{
  Internal function not to be called by the user
}
\details{
  Internal function. \code{get_effects} is called by
  \code{plspm}.
}
\section{Warning}{
  Do NOT use this function unless you are ME, a package
  developer, or a jedi user who really knows what is doing
  (seriously!)
}
\keyword{internal}

