\name{get_gof}
\alias{get_gof}
\title{Goodness-of-fit for \code{plspm}}
\usage{
  get_gof(comu, R2, blocks, IDM)
}
\arguments{
  \item{comu}{communalities}

  \item{R2}{R-squared coefficient}

  \item{blocks}{list of variables in each block}

  \item{IDM}{Inner Design Matrix}
}
\description{
  Internal function. \code{get_gof} is called by
  \code{plspm}
}
\keyword{internal}

