\name{get_GQI}
\alias{get_GQI}
\title{Group Quality Index}
\usage{
  get_GQI(pls, part, DM)
}
\arguments{
  \item{pls}{an object of class \code{"plspm"}}

  \item{part}{vector with unit memberships or categorical
  variable}

  \item{DM}{data matrix}
}
\description{
  Calculate Group Quality Index for REBUS-PLS
}
\keyword{internal}

