\name{get_locals_test}
\alias{get_locals_test}
\title{Local groups comparison test}
\usage{
  get_locals_test(X, pls, g)
}
\arguments{
  \item{X}{data matrix}

  \item{pls}{an object of class \code{"plspm"}}

  \item{g}{factor with 2 levels indicating the groups to be
  compared}
}
\value{
  list with path coefficients, loadings, and gof
}
\description{
  Calculate Group Quality Index for REBUS-PLS Internal
  function.
}
\keyword{internal}

