\name{get_path_scheme}
\alias{get_path_scheme}
\title{Calculate inner weighting path scheme}
\usage{
  get_path_scheme(IDM, Y)
}
\arguments{
  \item{IDM}{Inner Design Matrix}

  \item{Y}{Matrix of latent variables}
}
\description{
  Internal function. \code{get_path_scheme} is called by
  \code{get_weights}
}
\keyword{internal}

