\name{get_PLSR}
\alias{get_PLSR}
\title{Internal PLS regression (full data)}
\usage{
  get_PLSR(Y, X, ncomp)
}
\arguments{
  \item{Y}{an (already centered) vector of order n}

  \item{X}{an (already centered) n x p matrix}

  \item{ncomp}{number of components}
}
\description{
  Internal function not to be called by the user
}
\details{
  Internal function. \code{get_PLSR}
}
\section{Warning}{
  Do NOT use this function unless you are ME, a package
  developer, or a jedi user who really knows what is doing
  (seriously!)
}
\keyword{internal}

