\name{get_PLSRdoubleQ}
\alias{get_PLSRdoubleQ}
\title{get_PLSRdoubleQ}
\usage{
  get_PLSRdoubleQ(Y = NULL, Yc = NULL, X = NULL, Xc = NULL,
    ncomp)
}
\arguments{
  \item{Y}{numeric matrix}

  \item{Yc}{numeric matrix}

  \item{X}{numeric matrix}

  \item{Xc}{numeric matrix}

  \item{ncomp}{number of components}
}
\description{
  Internal function not to be called by the user
}
\details{
  Internal function. \code{get_PLSRdoubleQ}
}
\section{Warning}{
  Do NOT use this function unless you are ME, a package
  developer, or a jedi user who really knows what is doing
  (seriously!)
}
\keyword{internal}

