\name{normalize}
\alias{normalize}
\title{Normalize a vector}
\usage{
  normalize(x)
}
\arguments{
  \item{x}{a numeric vector}
}
\value{
  normalized vector
}
\description{
  Internal function not to be called by the user
}
\details{
  Internal function. \code{normalize} is called internally
}
\section{Warning}{
  Do NOT use this function unless you are ME, a package
  developer, or a jedi user who really knows what is doing
  (seriously!)
}
\keyword{internal}

