\name{plspm-internal}
\title{Internal Design Functions}
\alias{plsr1}
\alias{pls.weights}
\alias{pls.paths}
\alias{pls.efects}
\alias{pls.loads}
\alias{pls.unidim}
\alias{pls.boot}
\alias{pls.gof}
\alias{GQI}
\alias{locals.test}
\description{Internal Design functions.}
\details{These functions are not to be called by the user or are undocumented.}
\keyword{internal}

