% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm-package.R
\docType{package}
\name{plspm-package}
\alias{plspm-package}
\title{plspm: Partial Least Squares Path Modeling (PLS-PM)}
\description{
Partial Least Squares Path Modeling (PLS-PM), Tenenhaus, Esposito Vinzi, Chatelin, Lauro (2005) \doi{10.1016/j.csda.2004.03.005}, analysis for both metric and non-metric data, as well as REBUS analysis, Esposito Vinzi, Trinchera, Squillacciotti, and Tenenhaus (2008) \doi{10.1002/asmb.728}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gastonstat/plspm}
  \item Report bugs at \url{https://github.com/gastonstat/plspm/issues}
}

}
\author{
This package was written by Gaston Sanchez. 
Maintainer: Frederic Bertrand <frederic.bertrand@lecnam.net>
}
\keyword{internal}
