\name{print.local.models}
\alias{print.local.models}
\title{ Printing local.models objects }
\description{
  Printing method for objects of class \code{"local.models"}
}
\usage{
     \method{print}{local.models}(x, \dots)
}
\arguments{
  \item{x}{Object of class \code{"local.models"} returned by \code{\link{local.models}}.}
  \item{\dots}{Further arguments are ignored.}
}
\details{
  The function \code{local.models} displays the elements of a list containing all the results. 
}
\seealso{\code{\link{local.models}}}