\name{print.nipals}
\alias{print.nipals}
\title{Printing nipals obejcts}
\description{
  Printing method for objects of class \code{"nipals"}
}
\usage{
  \method{print}{nipals}(x, \dots)
}
\arguments{
  \item{x}{Object of class \code{"nipals"} returned by \code{\link{nipals}}.}
  \item{\dots}{Further arguments are ignored.}
}
\details{
  The function \code{print.nipals} displays the elements of a list containing all the results. 
}
\seealso{\code{\link{nipals}},\code{\link{plot.nipals}}}
