\name{print.plsca}
\alias{print.plsca}
\title{Printing plsca objects}
\description{
  Printing method for objects of class \code{"plsca"}
}
\usage{
  \method{print}{plsca}(x, \dots)
}
\arguments{
  \item{x}{Object of class \code{"plsca"} returned by \code{\link{plsca}}.}
  \item{\dots}{Further arguments are ignored.}
}
\details{
  The function \code{print.plsca} displays the elements of a list containing all the results. 
}
\seealso{\code{\link{plsca}},\code{\link{plot.plsca}}}
