\name{print.plspm.groups}
\alias{print.plspm.groups}
\title{Printing plspm.groups objects}
\description{
  Printing method for objects of class \code{"plspm.groups"}
}
\usage{
  \method{print}{plspm.groups}(x, \dots)
}
\arguments{
  \item{x}{Object of class \code{"plspm.groups"} returned by \code{\link{plspm.groups}}.}
  \item{\dots}{Further arguments are ignored.}
}
\details{
  The function \code{print.plspm.groups} displays a list with the results of the test. 
}
\seealso{\code{\link{plspm.groups}}, \code{\link{plspm}}}
