\name{print.plspm}
\alias{print.plspm}
\title{Printing plspm objects}
\description{
  Print method for objects of class \code{"plspm"}
}
\usage{
  \method{print}{plspm}(x, \dots)
}
\arguments{
  \item{x}{Object of class \code{"plspm"} returned by \code{\link{plspm}}.}
  \item{\dots}{Further arguments are ignored.}
}
\details{
  The function \code{print.plspm} displays the elements of a list containing all the results. 
}
\seealso{\code{\link{plspm}}}
