\name{print.plsreg1}
\alias{print.plsreg1}
\title{Printing plsreg1 objects}
\description{
  Printing method for objects of class \code{"plsreg1"}
}
\usage{
  \method{print}{plsreg1}(x, \dots)
}
\arguments{
  \item{x}{Object of class \code{"plsreg1"} returned by \code{\link{plsreg1}}.}
  \item{\dots}{Further arguments are ignored.}
}
\details{
  The function \code{print.plsreg1} displays the elements of a list containing all the results. 
}
\seealso{\code{\link{plsreg1}},\code{\link{plot.plsreg1}}}
