\name{print.plsreg2}
\alias{print.plsreg2}
\title{Printing plsreg2 objects}
\description{
  Printing method for objects of class \code{"plsreg2"}
}
\usage{
  \method{print}{plsreg2}(x, \dots)
}
\arguments{
  \item{x}{Object of class \code{"plsreg2"} returned by \code{\link{plsreg2}}.}
  \item{\dots}{Further arguments are ignored.}
}
\details{
  The function \code{print.plsreg2} displays the elements of a list containing all the results. 
}
\seealso{\code{\link{plsreg2}}, \code{\link{plot.plsreg2}}}
