\name{print.rebus}
\alias{print.rebus}
\title{Printing rebus objects}
\description{
  Printing method for objects of class \code{"rebus"}
}
\usage{
  \method{print}{rebus}(x, \dots)
}
\arguments{
  \item{x}{Object of class \code{"rebus"} returned by either \code{\link{it.reb}} or \code{\link{rebus.pls}}.}
  \item{\dots}{Further arguments are ignored.}
}
\details{
  The function \code{print.rebus} displays the elements of a list containing all the results. 
}
\seealso{\code{\link{rebus.pls}}}