\name{print.rebus.test}
\alias{print.rebus.test}
\title{Printing rebus.test objects}
\description{
  Printing method for objects of class \code{"rebus.test"}
}
\usage{
  \method{print}{rebus.test}(x, \dots)
}
\arguments{
  \item{x}{Object of class \code{"rebus.test"} returned by \code{\link{rebus.test}}. }
  \item{\dots}{Further arguments are ignored.}
}
\details{
  The function \code{print.rebus.test} displays the names of the elements from a list containing all the results for each pair of groups comparison. 
}
\seealso{\code{\link{rebus.test}}, \code{\link{rebus.pls}}}
