\name{print.summary.plspm}
\alias{print.summary.plspm}
\title{Printing summaries of plspm objects}
\description{
  Print method of summary for objects of class \code{"plspm"}
}
\usage{
 \method{print}{summary.plspm}(x, \dots)
}
\arguments{
  \item{x}{Object of class \code{"summary.plspm"}.}
  \item{\dots}{Further arguments are ignored.}
}
\seealso{\code{\link{plspm}}, \code{\link{summary.plspm}}.}
