\name{get_effects}
\alias{get_effects}
\title{Path coefficient effects for \code{plspm}}
\usage{
  get_effects(Path)
}
\arguments{
  \item{Path}{Matrix of path coefficients}
}
\value{
  data frame with direct, indirect, and total effects
}
\description{
  Internal function. \code{get_effects} is called by
  \code{plspm}.
}
\keyword{internal}

