\name{get_path_scheme}
\alias{get_path_scheme}
\title{Calculate inner weighting path scheme}
\usage{
  get_path_scheme(path_matrix, LV)
}
\arguments{
  \item{path_matrix}{matrix with paths}

  \item{LV}{matrix of latent variables}
}
\description{
  Internal function not to be called by the user
}
\details{
  Internal function. \code{get_path_scheme} is called by
  \code{get_weights}
}
\section{Warning}{
  Do NOT use this function unless you are ME, a package
  developer, or a jedi user who really knows what is doing
  (seriously!)
}
\keyword{internal}

