\name{plspm-package}
\alias{plspm-package}
\docType{package}
\title{
  Package of Partial Least Squares (PLS) data analysis methods.
}
\description{
  Contains functions of various Partial Least Squares methods such as the NIPALS algorithm, PLS Regression (PLSR1 and PLSR2), PLS Canonical Analysis, and PLS Path Modeling.
}
\details{
  \tabular{ll}{
  Package: \tab plspm\cr
  Type: \tab Package\cr
  Version: \tab 0.1-2\cr
  Date: \tab 2009-05-08\cr
  License: \tab GPL (>= 2)\cr
  LazyLoad: \tab yes\cr
  }
  The main functions are \code{\link{plspm}}, \code{\link{plspm.groups}}, \code{\link{plsreg1}}, \code{\link{plsreg2}}, \code{\link{plsca}}, and \code{\link{nipals}}. 
}
\author{
  Author: Gaston Sanchez \email{gaston.sanchez@upc.edu}

  Maintainer: Gaston Sanchez \email{plspmr@gmail.com}
}
\references{
  \url{http://www.docstoc.com/profile/gasigiri}
}
\keyword{ package }