\name{plspm-package}
\alias{plspm-package}
\docType{package}
\title{
  Package of partial least squares path modeling.
}
\description{
 Contains the functions to calculate path models with latent variables by partial least squares approach.
}
\details{
  \tabular{ll}{
  Package: \tab plspm\cr
  Type: \tab Package\cr
  Version: \tab 0.1-0\cr
  Date: \tab 2009-04-06\cr
  License: \tab GPL (>= 2)\cr
  }
  The main function is \code{\link{plspm}}. See help files for more details.
}
\author{
  Author: Gaston Sanchez \email{gaston.sanchez@upc.edu}

  Maintainer: Gaston Sanchez \email{plspmr@gmail.com}
}
