\name{plspm-package}
\alias{plspm-package}
\docType{package}
\title{
  Package of Partial Least Squares (PLS) data analysis methods.
}
\description{
  Contains functions of various Partial Least Squares methods such as the NIPALS algorithm, PLS Regression (PLSR1 and PLSR2), PLS Canonical Analysis, PLS Path Modeling, and REBUS-PLS.
}
\details{
  \tabular{ll}{
  Package: \tab plspm\cr
  Type: \tab Package\cr
  Version: \tab 0.2-0\cr
  Date: \tab 2011-12-18\cr
  Depends: \tab amap, diagram\cr
  Suggests: \tab FactoMineR\cr
  License: \tab GPL (>= 2)\cr
  LazyLoad: \tab yes\cr
  }
  The main functions are \code{\link{plspm}}, \code{\link{plspm.groups}}, \code{\link{rebus.pls}}, \code{\link{rebus.test}}, \code{\link{plsreg1}}, \code{\link{plsreg2}}, \code{\link{plsca}}, and \code{\link{nipals}}. 
}
\author{
  Author: Gaston Sanchez \email{plspmr@gmail.com}

  Maintainer: Gaston Sanchez \email{plspmr@gmail.com}
}
\references{
  \url{http://www.scribd.com/plspm}

  Tenenhaus, M. (1998) \emph{La Regression PLS. Theorie et Pratique}. Editions TECHNIP, Paris.

  \emph{Handbook of Partial Least Squares} (2010). Editors: V. Esposito Vinzo, W.W. Chin, J. Henseler, H. Wang. Springer, Paris.
}
\keyword{ package }