\name{plspm-package}
\alias{plspm-package}
\docType{package}
\title{
  Package of Partial Least Squares (PLS) data analysis methods.
}
\description{
  Contains functions of various Partial Least Squares methods such as the NIPALS algorithm, PLS Regression (PLSR1 and PLSR2), PLS Canonical Analysis, PLS Path Modeling, and REBUS-PLS.
}
\details{
  \tabular{ll}{
  Package: \tab plspm\cr
  Type: \tab Package\cr
  Version: \tab 0.1-5\cr
  Date: \tab 2010-18-05\cr
  License: \tab GPL (>= 2)\cr
  LazyLoad: \tab yes\cr
  }
  The main functions are \code{\link{plspm}}, \code{\link{plspm.groups}}, \code{\link{rebus.pls}}, \code{\link{rebus.test}}, \code{\link{plsreg1}}, \code{\link{plsreg2}}, \code{\link{plsca}}, and \code{\link{nipals}}. 
}
\author{
  Authors: Gaston Sanchez \email{gaston.sanchez@upc.edu}, and  Laura Trinchera \email{laura.trinchera@supelec.fr}

  Maintainer: Gaston Sanchez \email{plspmr@gmail.com}
}
\references{
  \url{http://www.docstoc.com/profile/gasigiri}

  \url{http://www.fedoa.unina.it/view/people/Trinchera,_Laura.html}
}
\keyword{ package }