\name{plspm-package}
\alias{plspm-package}
\docType{package}
\title{
  Tools for Partial Least Squares Path Modeling
}
\description{
  plspm contains a set of functions for performing Partial Least Squares Path Modeling (PLS-PM) analysis, as well as group comparison, and REBUS-PLS. This is a new version that no longer supports other PLS methods such as nipals, regression, and canonical analysis. For the later methods use the package plsdepot.
}
\details{
  \tabular{ll}{
  Package: \tab plspm\cr
  Type: \tab Package\cr
  Version: \tab 0.3.7\cr
  Date: \tab 2013-01-28\cr
  Depends: \tab R (>=2.14.0) amap, diagram\cr
  Suggests: \tab FactoMineR\cr
  License: \tab GPL-3\cr
  LazyLoad: \tab yes\cr
  } 
}
\author{
  Author: Gaston Sanchez, Laura Trinchera

  Maintainer: Gaston Sanchez \email{gaston.stat@gmail.com}
}
\references{
  \url{http://www.gastonsanchez.com/plspm}

  Sanchez G (2013) \emph{PLS Path Modeling with R} \cr
  \url{http://www.gastonsanchez.com/PLS_Path_Modeling_with_R.pdf}

  Tenenhaus M (1998) \emph{La Regression PLS. Theorie et Pratique}. Editions TECHNIP, Paris.

  \emph{Handbook of Partial Least Squares} (2010). Editors: V. Esposito Vinzo, W.W. Chin, J. Henseler, H. Wang. Springer, Paris.
}
\keyword{ package }