\name{plspm-package}
\alias{plspm-package}
\docType{package}
\title{
  Package of Partial Least Squares (PLS) data analysis methods.
}
\description{
  Contains functions of various Partial Least Squares methods such as the NIPALS algorithm, PLS Regression (PLSR1 and PLSR2), PLS Canonical Analysis, PLS Path Modeling, and REBUS-PLS.
}
\details{
  \tabular{ll}{
  Package: \tab plspm\cr
  Type: \tab Package\cr
  Version: \tab 0.1-11\cr
  Date: \tab 2010-08-09\cr
  Depends: \tab amap, diagram\cr
  Suggests: \tab FactoMineR\cr
  License: \tab GPL (>= 2)\cr
  LazyLoad: \tab yes\cr
  }
  The main functions are \code{\link{plspm}}, \code{\link{plspm.groups}}, \code{\link{rebus.pls}}, \code{\link{rebus.test}}, \code{\link{plsreg1}}, \code{\link{plsreg2}}, \code{\link{plsca}}, and \code{\link{nipals}}. 
}
\author{
  Authors: Gaston Sanchez \email{plspmr@gmail.com}, and  Laura Trinchera \email{laura.trinchera@supelec.fr}

  Maintainer: Gaston Sanchez \email{plspmr@gmail.com}
}
\references{
  \url{http://www.scribd.com/plspm}

  \url{http://www.fedoa.unina.it/view/people/Trinchera,_Laura.html}

  Esposito Vinzi, V., Chin, W.W., Henseler, J., Wang, H. (eds) (2010) \emph{Handbook of Partial Least Squares.} Springer-Verlag Berlin Heidelberg.
}
\keyword{ package }